+++ app/Http/Controllers/api/v1/PermissionController.php
@@ -3,19 +3,18 @@
 namespace App\Http\Controllers\api\v1;
 
 use App\Http\Controllers\Controller;
-use App\Http\Resources\Permission as PermissionResource;
+use App\Http\Resources\PermissionResourceCollection;
 use App\Permission;
-use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 
 class PermissionController extends Controller
 {
     /**
      * Display a listing of the resource.
      *
-     * @return AnonymousResourceCollection
+     * @return PermissionResourceCollection
      */
     public function index()
     {
-        return PermissionResource::collection(Permission::all());
+        return new PermissionResourceCollection(Permission::all());
     }
 }