+++ resources/js/policies/index.js
@@ -1,4 +1,5 @@
 import SettingPolicy from './SettingPolicy';
+import ApplicationSettingPolicy from './ApplicationSettingPolicy';
 import RolePolicy from './RolePolicy';
 import RoomPolicy from './RoomPolicy';
 import UserPolicy from './UserPolicy';
@@ -8,6 +9,7 @@ import ServerPoolPolicy from './ServerPoolPolicy';
 
 export default {
   SettingPolicy,
+  ApplicationSettingPolicy,
   RolePolicy,
   RoomPolicy,
   UserPolicy,