+++ resources/js/views/settings/Application.vue
@@ -1,9 +1,6 @@
 <template>
-  <b-container fluid>
-    <h3>
-      {{ $t('settings.application.title') }}
-    </h3>
-
+ <div>
+    <h3>{{ $t('settings.application.title') }}</h3>
     <hr>
     <b-overlay :show="isBusy || !loaded">
 
@@ -541,7 +538,7 @@
         </div>
       </b-form>
     </b-overlay>
-  </b-container>
+ </div>
 </template>
 
 <script>
@@ -722,7 +719,7 @@ export default {
      * Check if user is only allowed to read settings
      */
     viewOnly () {
-      return PermissionService.cannot('update', 'SettingPolicy');
+      return PermissionService.cannot('update', 'ApplicationSettingPolicy');
     },
 
     /**